2nd commit
Added class Rectangle for carriage and ball
