+++ Arkanoid.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30413.136
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "gems", "gems.vcxproj", "{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "arkanoid", "arkanoid.vcxproj", "{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,19 +13,19 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Debug|x64.ActiveCfg = Debug|x64
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Debug|x64.Build.0 = Debug|x64
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Debug|x86.ActiveCfg = Debug|Win32
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Debug|x86.Build.0 = Debug|Win32
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Release|x64.ActiveCfg = Release|x64
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Release|x64.Build.0 = Release|x64
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Release|x86.ActiveCfg = Release|Win32
-		{E2AB60A5-49CA-48A1-A35D-5641F3C1ACA2}.Release|x86.Build.0 = Release|Win32
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Debug|x64.ActiveCfg = Debug|x64
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Debug|x64.Build.0 = Debug|x64
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Debug|x86.ActiveCfg = Debug|Win32
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Debug|x86.Build.0 = Debug|Win32
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Release|x64.ActiveCfg = Release|x64
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Release|x64.Build.0 = Release|x64
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Release|x86.ActiveCfg = Release|Win32
+		{2DA1E954-EAE9-4DFF-8E0A-35D13CA39660}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {6497A1F8-F23F-4E68-ABD2-E8C272D3D079}
+		SolutionGuid = {6736327C-B333-4024-94E1-25B4EB8D82B7}
 	EndGlobalSection
 EndGlobal
