--- Arkanoid/README.md
@@ -1,9 +1,0 @@
-# my_cplusplus_labs
-
-One branch is for one lab:
-
-1st branch - Quadratic Equation
-
-2nd - Math Exam (using 'Quadratic Equation')
-
-3rd - Arkanoid game
